refactor(middleware): split config-input types from the API-surface types
Move KyaOsConfig/KyaOsDelegationConfig/KyaOsIdentityConfig/KYA_OS_ACTIONS into with-kya-os.config-types.ts, re-exported from with-kya-os.types.ts so importers are unchanged. Both files now <300. Completes the with-kya-os.ts split: 1649 -> 182 across focused modules, all <300; 1607 tests green. Phase 2e.
