+++ src/middleware/with-kya-os.types.ts
@@ -1,160 +1,24 @@
 /**
- * KYA-OS Middleware — public type surface.
+ * KYA-OS Middleware — public API-surface types.
  *
- * The config, tool, and role interfaces for `createKyaOsMiddleware` / `withKyaOs`,
- * separated from the implementation (`./with-kya-os.ts`) so the ~300-line type
- * surface does not inflate the factory module. Re-exported verbatim from
- * `./with-kya-os.ts`, so every existing `import { KyaOsConfig } from './with-kya-os'`
+ * The tool, call-context, and role interfaces for `createKyaOsMiddleware` /
+ * `withKyaOs`, separated from the implementation (`./with-kya-os.ts`). The
+ * configuration INPUT types live in `./with-kya-os.config-types.ts` and are
+ * re-exported below, so every existing `import { KyaOsConfig } from './with-kya-os'`
  * is unchanged.
  */
 
-import type { FetchProvider, NonceCacheProvider } from "../providers/base.js";
-import type { AuditLogProvider } from "../providers/audit-log.js";
-import type { GrantStore } from "../providers/grant-store.js";
-import type { SessionManager, SessionConfig } from "../session/manager.js";
+import type { SessionManager } from "../session/manager.js";
 import type { ProofGenerator } from "../proof/generator.js";
-import type {
-  DIDResolver,
-  StatusListResolver,
-} from "../delegation/vc-verifier.js";
-import type { DIDResolverRegistry } from "../delegation/did-resolver-registry.js";
-import type { RevocationChecker } from "../delegation/chain-enforcement.js";
-import type {
-  DelegationCredential,
-  NeedsAuthorizationError,
-} from "../types/protocol.js";
+import type { NeedsAuthorizationError } from "../types/protocol.js";
 import type { PolicyEngine } from "../policy/engine.js";
 import type { RiskClassifier } from "../policy/classifier.js";
 import type { ApprovalGrant } from "../policy/approval.js";
+import type { KyaOsIdentityConfig } from "./with-kya-os.config-types.js";
 
-export interface KyaOsIdentityConfig {
-  did: string;
-  kid: string;
-  privateKey: string;
-  publicKey: string;
-  agentName?: string;
-}
-
-export const KYA_OS_ACTIONS = ["handshake", "identity", "reputation"] as const;
-export type KyaOsAction = (typeof KYA_OS_ACTIONS)[number];
-
-export interface KyaOsDelegationConfig {
-  /**
-   * Optional custom DID resolver. If it returns null, middleware falls back to
-   * built-in did:key resolution and fetch-backed did:web resolution.
-   */
-  didResolver?: DIDResolver;
-  /**
-   * Optional fetch provider used for did:web resolution.
-   * If omitted, middleware falls back to the runtime global fetch when available.
-   */
-  fetchProvider?: FetchProvider;
-  /**
-   * Optional DID method resolver registry. Entries are keyed by DID method
-   * (for example, "cheqd") and are checked before the built-in did:key and
-   * did:web fallback. Factory entries receive the active fetch provider.
-   */
-  didResolvers?: DIDResolverRegistry;
-  /**
-   * Resolver for StatusList2021 checks. Credentials with credentialStatus are
-   * rejected when no resolver is configured.
-   */
-  statusListResolver?: StatusListResolver;
-  /**
-   * Resolve ancestor credentials for a delegated chain. The returned array may
-   * contain only ancestors (root -> parent) or the full chain (root -> leaf).
-   */
-  resolveDelegationChain?: (
-    leafCredential: DelegationCredential,
-  ) => Promise<DelegationCredential[]>;
-  /**
-   * Graph-backed ancestor-revocation check. When provided, the chain walk also
-   * verifies the leaf is not revoked via a cascade-revoked ANCESTOR — caught
-   * independently of how `resolveDelegationChain` hydrated the chain (it walks
-   * the delegation graph + StatusList). `CascadingRevocationManager` is the
-   * reference adapter. Omit to keep the prior per-credential-status behavior.
-   */
-  revocationChecker?: RevocationChecker;
-  /**
-   * Holder-of-key enforcement for inbound calls (spec §11.8). A valid delegation
-   * is a *bearer* credential; holder binding additionally requires the caller to
-   * present a per-request proof (`_kyaos_proof`) signed by the delegation
-   * SUBJECT's key, closing theft-replay for the key-bearing population.
-   *
-   * - `'off'` (default): no holder-binding check — current behavior, unchanged.
-   * - `'warn'`: check and log failures (missing/unbound proof), but still allow.
-   * - `'enforce'`: reject calls whose proof is missing or does not bind the
-   *   subject. **Breaking for callers that don't yet send `_kyaos_proof`** — opt
-   *   in only once your agents mint request proofs.
-   *
-   * Scope is did:key subjects (the DID encodes the key). did:web and other
-   * subjects are deferred to cnf-based binding (phase 2) and logged, never
-   * rejected — so enabling `'enforce'` never breaks did:web traffic.
-   */
-  holderBinding?: "off" | "warn" | "enforce";
-}
-
-export interface KyaOsConfig {
-  /** Agent identity (DID + key material) */
-  identity: KyaOsIdentityConfig;
-  /** Session configuration overrides */
-  session?: Omit<SessionConfig, "nonceCache">;
-  /**
-   * Replay-protection store, shared by the session handshake AND inbound holder
-   * binding so both draw on one nonce namespace. Defaults to an in-memory store
-   * (single-process only); inject a Redis / Durable Object / KV-backed
-   * {@link NonceCacheProvider} for multi-instance deployments. Set here, not on
-   * `session`, so there is exactly one cache and the two cannot diverge.
-   */
-  nonceCache?: NonceCacheProvider;
-  /**
-   * Durable store for approved authorization grants, enabling the no-paste
-   * retry: an agent that already obtained a grant can call again — even on a
-   * fresh instance with empty memory, or after a restart — without re-pasting
-   * the delegation. Defaults to an in-memory {@link MemoryGrantStore}
-   * (single-process only); inject a Redis / Durable Object / DB-backed
-   * {@link GrantStore} for multi-instance deployments. Mirrors `nonceCache?`.
-   *
-   * For the no-paste retry to actually RESOLVE a bound grant, the call must be
-   * resolvable: either holder binding is `'enforce'` and the agent presents a
-   * per-request `_kyaos_proof` (agent-anchored `getByAgent`), or the client
-   * threads its `sessionId` (session-bearer `getBySession`). A grant that would
-   * be unresolvable — holder binding `'off'` with no threaded `sessionId` — is
-   * NOT bound (it would only orphan a store row); such flows achieve durability
-   * by re-presenting the delegation instead.
-   */
-  grantStore?: GrantStore;
-  /**
-   * Whether to ALSO emit the detached proof under the legacy bare `_meta.proof`
-   * key (in addition to the namespaced `org.kya-os/proof`). The value is
-   * identical under both keys and `_meta` is excluded from the response hash
-   * (§7.6), so the mirror never affects signatures or verification.
-   *
-   * Default `true` — a transition aid so pre-1.1 clients that still read bare
-   * `proof` keep working. Set `false` once your clients read the namespaced key
-   * (the examples do this for a clean single-key Inspector view).
-   *
-   * Stays ON by default for the whole 1.x line (a pre-1.1 reader of bare
-   * `_meta.proof` would otherwise silently get no proof); the legacy mirror is
-   * dropped at 2.0.
-   */
-  emitLegacyProofKey?: boolean;
-  /** Delegation verification overrides */
-  delegation?: KyaOsDelegationConfig;
-  /**
-   * When true, automatically creates a session on the first tool call
-   * if no session exists. Useful for demos and development where
-   * MCP clients don't support the `_kyaos` handshake flow.
-   * In production, KYA-OS-aware runtimes should execute handshake before tool calls.
-   */
-  autoSession?: boolean;
-  /**
-   * Sink for retaining audit records of verified tool calls. Defaults to a
-   * no-op; supply an {@link AuditLogProvider} (e.g. a durable, append-only
-   * implementation) to capture an audit trail.
-   */
-  auditLog?: AuditLogProvider;
-}
+// The configuration input types (KyaOsConfig, KyaOsDelegationConfig, …) live in
+// ./with-kya-os.config-types.ts; re-export them so importers are unchanged.
+export * from "./with-kya-os.config-types.js";
 
 export interface KyaOsToolDefinition {
   name: string;
